Change int to enum

+++ src/main/java/com/avalanche/tmcs/job_posting/JobPostingController.java
@@ -143,17 +143,17 @@ public class JobPostingController {
         switch (status) {
         case "active":
         	int duration = jobPosting.getDuration();
-        	jobPosting.setStatus(JobPosting.Status.ACTIVE.toInt());
+        	jobPosting.setStatus(JobPosting.Status.ACTIVE);
         	jobPosting.setNumDaysRemaining(duration);
             break;
         case "inactive":
-        	jobPosting.setStatus(JobPosting.Status.INACTIVE.toInt());
+        	jobPosting.setStatus(JobPosting.Status.INACTIVE);
             break;
         case "archived":
-        	jobPosting.setStatus(JobPosting.Status.ARCHIVED.toInt());
+        	jobPosting.setStatus(JobPosting.Status.ARCHIVED);
             break;
         case "needs_detailing":
-        	jobPosting.setStatus(JobPosting.Status.NEEDS_DETAILING.toInt());
+        	jobPosting.setStatus(JobPosting.Status.NEEDS_DETAILING);
             break;
         default:
         	return ResponseEntity.badRequest().build();
@@ -168,7 +168,7 @@ public class JobPostingController {
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public void deleteJobPosting(@PathVariable long id){
         JobPosting toDelete = jobPostingDAO.findOne(id);
-        toDelete.setStatus(JobPosting.Status.ARCHIVED.toInt());
+        toDelete.setStatus(JobPosting.Status.ARCHIVED);
         jobPostingDAO.save(toDelete);
     }
 
@@ -196,16 +196,16 @@ public class JobPostingController {
         List<JobPosting> jobPostings;
         switch (status) {
         case "active":
-        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.ACTIVE.toInt());
+        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.ACTIVE);
             break;
         case "inactive":
-        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.INACTIVE.toInt());
+        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.INACTIVE);
             break;
         case "archived":
-        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.ARCHIVED.toInt());
+        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.ARCHIVED);
             break;
         case "needs_detailing":
-        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.NEEDS_DETAILING.toInt());
+        	jobPostings = jobPostingDAO.findAllByCompanyAndStatus(companyWithID, JobPosting.Status.NEEDS_DETAILING);
             break;
         default:
             jobPostings = jobPostingDAO.findAllByCompany(companyWithID);
@@ -223,7 +223,7 @@ public class JobPostingController {
 
         List<JobPosting> jobPostings = jobPostingDAO.findAllByCompanyAndStatus(
                 companyWithID,
-                JobPosting.Status.ACTIVE.toInt()
+                JobPosting.Status.ACTIVE
         );
 
         return ResponseEntity.ok(jobPostings);
@@ -239,7 +239,7 @@ public class JobPostingController {
 
         List<JobPosting> jobPostings = jobPostingDAO.findAllByCompanyAndStatus(
                 companyWithID,
-                JobPosting.Status.INACTIVE.toInt()
+                JobPosting.Status.INACTIVE
         );
 
         return ResponseEntity.ok(jobPostings);
@@ -255,7 +255,7 @@ public class JobPostingController {
 
         List<JobPosting> jobPostings = jobPostingDAO.findAllByCompanyAndStatus(
                 companyWithID,
-                JobPosting.Status.ARCHIVED.toInt()
+                JobPosting.Status.ARCHIVED
         );
 
         return ResponseEntity.ok(jobPostings);
@@ -280,7 +280,7 @@ public class JobPostingController {
     @RequestMapping(value = "/{id}/fulfill", method = RequestMethod.POST)
     public void fulfillJobPosting(@PathVariable long id){
         JobPosting toFulfill = jobPostingDAO.findOne(id);
-        toFulfill.setStatus(JobPosting.Status.INACTIVE.toInt());
+        toFulfill.setStatus(JobPosting.Status.INACTIVE);
         jobPostingDAO.save(toFulfill);
     }
